backend/backend/requirements.txt
Django
djangorestframework
mysql-connector-python
mysqlclient